backend/requirements.txt
celery>=5.2.0
Django==3.2.6
redis==3.5.3
psycopg2-binary==2.8.6
django-cors-headers
djangorestframework
